refactor(ot-40): extract DTOs to separate files and add class-validator decorators
