+++ api-gateway/src/task/dto/update-task.dto.ts
@@ -1,0 +1,19 @@
+import { IsNotEmpty, IsString, IsOptional, IsBoolean } from 'class-validator';
+
+export class UpdateTaskDto {
+    @IsNotEmpty()
+    @IsString()
+    title: string;
+
+    @IsString()
+    @IsOptional()
+    description?: string;
+
+    @IsBoolean()
+    @IsOptional()
+    isCompleted?: boolean;
+
+    @IsOptional()
+    @IsString()
+    dueDate?: string;
+}
